+++ config/config.go
@@ -153,6 +153,7 @@ func Init() error {
 	viper.SetDefault("a-general.downloads", "")
 	viper.SetDefault("a-general.page_max_size", 2097152)
 	viper.SetDefault("a-general.page_max_time", 10)
+	viper.SetDefault("a-general.emoji_favicons", false)
 	viper.SetDefault("cache.max_size", 0)
 	viper.SetDefault("cache.max_pages", 20)