+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Emoji favicons can now be seen if `emoji_favicons` is enabled in the config (#62)
 
 ### Changed
 
 ### Fixed
 - Two digit (and higher) link texts are now in line with one digit ones (#60)
+- Race condition when reloading pages, could have caused the cache to still be used
 
 
 ## [1.4.0] - 2020-07-28